+++ models/forms/CreateAccountForm.php
@@ -1,5 +1,5 @@
 <?php
-namespace app\models;
+namespace app\models\forms;
 
 use yii\base\Model;
